% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbranch.R
\name{compute_all_distances}
\alias{compute_all_distances}
\title{Compute sample distances}
\usage{
compute_all_distances(input_dat)
}
\arguments{
\item{input_dat:}{data frame of input data with rows=samles and cols=dimensions}
}
\value{
Dist: a matrix containing all sample distances (square Euclidean)
}
\description{
Computes all SQUARED pairwise distances of data points.
}
\examples{
\dontrun{
 Dist=compute_all_distances(scdata.3lines)
}
}

