% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbranch.R
\name{find_K_nearest}
\alias{find_K_nearest}
\title{Find K nearest neighbours}
\usage{
find_K_nearest(n, K = 3, Dist)
}
\arguments{
\item{n:}{sample for which the nearest neighbors are computed}

\item{K:}{number of nearest neighbors to be considered}

\item{Dist:}{matrix containing sample distances, initially computed by compute_all_distances}
}
\value{
a list with elements:
\itemize{
 \item - pos: the positions (rows) of the N nearest neighbours.
 \item - dst: the distances of the N nearest neighbours.
}
}
\description{
Finds the K nearest neighbours of a sample, given a matrix containing sample distances.
}
\examples{
\dontrun{
 Dist=compute_all_distances(scdata.3lines.simulated6genes_subsampled)

 #find the 5 nearest neghbours of the second sample
 neibs=find_K_nearest(n=2,K=5,Dist=Dist)
}
}

