% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-regions.R
\name{identify_regions}
\alias{identify_regions}
\title{Identify regions of interest}
\usage{
identify_regions(input_dat, mode = "tip", tip_mode = "3",
  gap_scores = NULL, smoothing_region = NULL,
  smoothing_region_thresh = NULL, Dist, dotsize1 = 7, dotsize2 = 7,
  nclust = NULL, nclust.max = 5, B.max = 100, SE.factor = 2.58,
  repeats = 20, gapm_tips = F)
}
\arguments{
\item{input_dat:}{data frame of input data with rows=samles and cols=dimensions.}

\item{mode:}{=c('tip','branch') find either the tips or the branching regions}

\item{tip_mode:}{=c('2','3','both') find the tips using the GAP statistic of 1 vs 2,3 or both}

\item{gap_scores:}{list, output of function 'kbranches.local' for the same 'input_dat'}

\item{smoothing_region:}{number of neighbours in to consider for label filtering}

\item{smoothing_region_thresh:}{minimum number of thresholds with same label in the neighbourhood required for the sample to keep it's label}

\item{Dist:}{matrix of sample to sample distances}

\item{dotsize1:}{size of points of class1 (red), used when plotting in 3D}

\item{dotsize2:}{size of points of class2 (green), used when plotting in 3D}

\item{nclust:}{number of clusters (tips/branching regions). If left NULL, will be estimated.}

\item{nclust.max:}{maximum possible number of clusters to consider. Only used if nclust==NULL}

\item{B.max:}{maximum number of bootstrap datasets used to calculate the GAP statistic to estimate nclust. Only used if nclust==NULL}

\item{SE.factor:}{used to estimate nclust, argument to cluster::maxSE. Only used if nclust==NULL}

\item{repeats:}{number of times to estimate nclust (for stability - the nclust most frequently considered 'best' is finally extracted). Only used if nclust==NULL}
}
\value{
a list with elements:
\itemize{
 \item - is_in_region: a logical vector indicating which samples are part of the region (TRUE) and which not (FALSE)
 \item - is_in_region_filtered: same as 'is_in_region', but after S_neib filtering to reduce noise.
 \item - cluster: cluster assignment for each of the data points in is_in_region_filtered
 \item - nclust: number of clusters identified, or specified by the user
 \item - cluster_frequency: how many times (out of 'repeats') each number of clusters (of the nclust.max) was considered 'best'.
 NULL if nclust was provided by the user
}
}
\description{
Identifies regions of interest (the tips of branches, or branching regions)
based on comparing the GAP scores#'acquired by running local clustering by 'kbranches.local'.
Performs local filtering to reduce noise in the extracted labels.
}
\examples{

#see example of kbranches.local

}

