% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbranch.R
\docType{package}
\name{kbranch-package}
\alias{kbranch}
\alias{kbranch-package}
\title{Clustering on single cell trajectories.}
\description{
The \pkg{kbranch} package provides two main functions
\itemize{
   \item \code{\link{kbranch.global}}
   \item \code{\link{kbranch.local}}
}
}

