% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbranches-package.R
\docType{package}
\name{kbranches-package}
\alias{kbranches}
\alias{kbranches-package}
\title{Clustering on single cell trajectories.}
\description{
The \pkg{kbranches} package provides two main functions
\itemize{
   \item \code{\link{kbranches.global}}
   \item \code{\link{kbranches.local}}
}
}

