% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refilter-identified-regions.R
\name{refilter_identified_regions}
\alias{refilter_identified_regions}
\title{Refilter already identified regions}
\usage{
refilter_identified_regions(input_dat, sample_labels = NULL,
  smoothing_region = NULL, smoothing_region_thresh = NULL, Dist,
  dotsize1 = 7, dotsize2 = 7)
}
\arguments{
\item{input_dat:}{data frame of input data with rows=samles and cols=dimensions.}

\item{sample_labels:}{vector of TRUE/FALSE labels, one for each sample. If sample_labels[i]==TRUE, then sample i is in region of interest (e.g. branching region)}

\item{smoothing_region:}{number of neighbours in to consider for label filtering}

\item{smoothing_region_thresh:}{minimum number of thresholds with same label in the neighbourhood required for the sample to keep it's label}

\item{Dist:}{matrix of sample to sample distances}

\item{dotsize1:}{size of points of class1 (red), used when plotting in 3D}

\item{dotsize2:}{size of points of class2 (green), used when plotting in 3D}
}
\value{
a list with elements:
\itemize{
 \item - is_in_region: a logical vector indicating which samples are part of the region (TRUE) and which not (FALSE)
 \item - is_in_region_filtered: same as 'is_in_region', but after S_neib filtering to reduce noise.
}
}
\description{
Refilters the results extracted by identify_regions().
Used only in rare cases where there is some noise remaining in the result of identify_regions().
}
\examples{

# TODO: example coming soon...

}

