% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-neighborhood-size.R
\name{scale_S_neib}
\alias{scale_S_neib}
\title{Scale the neighbourhood size}
\usage{
scale_S_neib(input_dat, n, K_init, Dist, dmin)
}
\arguments{
\item{input_dat:}{data frame of input data with rows=samles and cols=dimensions}

\item{n:}{position (row) of the sample, in the input_dat data frame}

\item{K_init:}{initial number of nearest neighbors that needs to be scaled}

\item{Dist:}{distance matrix of all samples}

\item{dmin:}{distance threshold (to the furthest neighbor) to be reached}
}
\value{
Kscaled: the scaled neighbourhood size
}
\description{
scales the number of neighbors (neighbourhood size) depending on the local sample density.
keeps increasing K until the distance to the furthest of the K neighbors is above the dmin threshold
}
\examples{
\dontrun{
Dist=compute_all_distances(scdata.3lines)

#find the 5 nearest neghbours of the second sample
neibs=find_K_nearest(n=2,K=5,Dist=Dist)
}
}

